requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.1.7
django-filter==22.1
django-phone-field==1.8.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
phonenumbers==8.13.7
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
